The vert.x gradle template was broken - specifically the Java PingVerticle had been amended not to deploy a listener on the event bus for "ping" messages but instead to deploy mod-webserver. This in turn caused the integration tests to fail due to timeout expecting a "pong" response (ModuleIntegrationTest in the groovy integration tests).
Furthermore the deployment of mod-webserver caused numerous "Address already in use" exceptions during the running of the unit tests. This deployment isn't tested by any test (as removing it and re-testing proved) so I wonder if the change which causes the deployment of mod-webserver was an accidental commit (the commit comment was "blah"). I've reinstated the previous version of the Java PingVerticle.